website/docs/r/cognito_user_pool.markdown: Document attribute constraints requirement
There is a known issue with the `aws_cognito_user_pool` resource that causes the recreation of schema attributes of type `String` or `Number` if the respective attribute constraints block is not defined. This change does not fix the issue, but updates the documentation for the resource with a note indicating the requirements to prevent future invocations of the said issue.